feat: add destructive_secondary style and extract destructive logic
- Extracted destructive logic from `src/model/command.rs` to `src/utils/destructive.rs`.
- Added `destructive_secondary` and `highlight_destructive_secondary` to theme config.
- Refactored `src/widgets/command.rs` to apply the new secondary styles for variables in destructive commands and simplified `unwrap_or` logic.
- Updated configuration docs and pre-existing vhs themes.
