+++ default_config.toml
@@ -161,6 +161,8 @@ comment = "italic green"
 error = "dark red"
 # The style used for destructive commands
 destructive = "red"
+# The style used for secondary parts of destructive commands
+destructive_secondary = "dim red"
 # The background color for the highlighted item in a list. Use "none" for no background color
 highlight = "darkgrey"
 # The string symbol displayed next to the highlighted item
@@ -175,6 +177,8 @@ highlight_accent = "yellow"
 highlight_comment = "italic green"
 # The destructive style applied specifically to a highlighted item
 highlight_destructive = "red"
+# The secondary destructive style applied specifically to a highlighted item
+highlight_destructive_secondary = "dim red"
 
 # --------------------------------------------------------------
 #   Search Tuning
