+++ vhs/themes/arcade.toml
@@ -4,9 +4,13 @@ secondary = "dim"
 accent = "136"
 comment = "rgb(106, 153, 66)"
 error = "dark red"
+destructive = "196"
+destructive_secondary = "dim 196"
 highlight = "none"
 highlight_symbol = "» "
 highlight_primary = "220"
 highlight_secondary = "222"
 highlight_accent = "208"
 highlight_comment = "rgb(143, 221, 75)"
+highlight_destructive = "196"
+highlight_destructive_secondary = "dim 196"
