+++ vhs/themes/dracula.toml
@@ -4,9 +4,13 @@ secondary = "#6272a4"
 accent = "bold #50fa7b"
 comment = "italic #bd93f9"
 error = "bold #ff5555"
+destructive = "bold #ff5555"
+destructive_secondary = "#ff5555"
 highlight = "#44475a"
 highlight_symbol = "◆ "
 highlight_primary = "#f8f8f2"
 highlight_secondary = "dim #f8f8f2"
 highlight_accent = "bold #ff79c6"
 highlight_comment = "italic #bd93f9"
+highlight_destructive = "bold #ff5555"
+highlight_destructive_secondary = "#ff5555"
